ProbModelChecker.clear(): deref complete var arrays, fix StateModelChecker constructor
Before, if varNum did not match the array length, some varDDRowVars/varDDColVars
would not be dereferenced.

Now, deref all variables in the arrays if they are non-null.

In the StateModelChecker constructor for the dummy model, we can now
pass the correct number of variables and a copy of the varDDRowVars array.



git-svn-id: https://www.prismmodelchecker.org/svn/prism/prism/trunk@10447 bbc10eb1-c90d-0410-af57-cb519fbb1720