lpsolve5j.cpp: Fix deletes
The Java wrapper around the C++ lpsolve library uses the construct
  delete a, b, c;
in several locations. The apparent goal is to delete all the pointers,
however the semantics of the C++ comma operator intervene: Only the
first pointer is deleted.

It looks like the relevant locations are never reached from PRISM code,
so we fix it mostly to silence the compiler warnings.


git-svn-id: https://www.prismmodelchecker.org/svn/prism/prism/trunk@10571 bbc10eb1-c90d-0410-af57-cb519fbb1720