arser.ExplicitFiles2ModulesFile: create label definitions in the ModulesFile for all labels in .lab file [with Steffen Maercker]
To be able to reference labels that are loaded from a .lab file and stored in the model,
they have to appear in the ModulesFile. This method stores

label "labelName" = false;

definitions. As the actual state set information is directly stored in the model,
evaluating the label will result in the correct state set.


git-svn-id: https://www.prismmodelchecker.org/svn/prism/prism/trunk@11740 bbc10eb1-c90d-0410-af57-cb519fbb1720