Cache mapping from original to product model in LUT
Previously, mapping a state from the original to the
product model required a linear search in the list
of inital states of the product model.
This causes quadratic time complexity when mapping
a result vector with the ModelTransformation methods.

To rectify this, we cache the mapping in a LUT.
Since the issue appears only in context of composed
model transformations, we initialize the LUT lazily.