Support for building interval models (IDTMCs and IMDPs).
In PRISM language models, transition probabilities in guards
can be defined as intervals, using the notation "[l,u]", where
l and u are expressions that can be evaluated as doubles.
There are no explicit idtmc or imdp model keywords; instead the
model type is inferred automatically for DTMCs or MDPs that
contain probabilities specified as intervals.

In the parser, ExpressionInterval is added to the AST classes.

The ModelGenerator interface has several new methods:
 * getIntervalEvaluator()
 * getTransitionProbabilityInterval()
 * getTransitionProbabilityObject()
 * getTransitionProbabilityString()
For the first, a new Evaluator for double intervals is added.

ModulesFileModelGenerator adds support for IDTMCs and IMDPs
specified in the PRISM language. This is done by making
the TransitionList and Update be of type either Value or
Interval<Value>, depending on the model type, but little else
is needed: generics and the new Evaluator take care of it.

For the explicit engine, we add new interfaces IDTMC and IMDP,
and implementations IDTMCSimple and IMDPSimple, which are
Mainly just wrappers around existing classes via generic types,
e.g., IDTMC<Value> extends DTMC<Interval<Value>>.